# Inventory-Service/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: BookYourShow - Inventory API
  description: REST APIs for managing inventory (movies, theatres, shows, seats) in BookYourShow Application
  version: 1.0.0
  contact:
    name: bookYourShow Team
    email: bookyourshow@company.com
...
servers:
  - url: http://localhost:8080/inventory/v1
    description: dev
paths:
...
  /cities:
    get:
      tags: [Cities]
      summary: Get all cities
      operationId: getCities
      parameters: []
      responses:
        '200':
          description: List of Cities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityListResponse'
        '500':
          description: Internal server error
    post:
      tags: [Cities]
      summary: Create a new city (admin)
      operationId: createCity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityCreateRequest'
      responses:
        '201':
          description: City created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: City already exists
  /cities/{id}:
    put:
      tags: [Cities]
      summary: Update a city (admin)
      operationId: updateCity
      parameters:
        - in: path
          name: id
          required: true
          description: City ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityUpdateRequest'
      responses:
        '200':
          description: City updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/City'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: City not found
    delete:
      tags: [Cities]
      summary: Delete a city (admin)
      operationId: deleteCity
      parameters:
        - in: path
          name: id
          required: true
          description: City ID
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: City deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: City not found
        '409':
          description: City still referenced by theatres

  # ── Languages ───────────────────────────────────────────
  /languages:
    get:
      tags: [Languages]
      summary: Get all languages
      operationId: getLanguages
      responses:
        '200':
          description: List of languages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguageListResponse'
        '500':
          description: Internal server error
    post:
      tags: [Languages]
      summary: Create a new language (admin)
      operationId: createLanguage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LanguageCreateRequest'
      responses:
        '201':
          description: Language created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Language already exists
  /languages/{id}:
    put:
      tags: [Languages]
      summary: Update a language (admin)
      operationId: updateLanguage
      parameters:
        - in: path
          name: id
          required: true
          description: Language ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LanguageUpdateRequest'
      responses:
        '200':
          description: Language updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Language not found
    delete:
      tags: [Languages]
      summary: Delete a language (admin)
      operationId: deleteLanguage
      parameters:
        - in: path
          name: id
          required: true
          description: Language ID
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: Language deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Language not found
        '409':
          description: Language still referenced by movies

  # ── Formats ─────────────────────────────────────────────
  /formats:
    get:
      tags: [Formats]
      summary: Get all formats
      operationId: getFormats
      responses:
        '200':
          description: List of formats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormatListResponse'
        '500':
          description: Internal server error
    post:
      tags: [Formats]
      summary: Create a new format (admin)
      operationId: createFormat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormatCreateRequest'
      responses:
        '201':
          description: Format created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Format'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Format already exists
  /formats/{id}:
    put:
      tags: [Formats]
      summary: Update a format (admin)
      operationId: updateFormat
      parameters:
        - in: path
          name: id
          required: true
          description: Format ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormatUpdateRequest'
      responses:
        '200':
          description: Format updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Format'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Format not found
    delete:
      tags: [Formats]
      summary: Delete a format (admin)
      operationId: deleteFormat
      parameters:
        - in: path
          name: id
          required: true
          description: Format ID
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: Format deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Format not found
        '409':
          description: Format still referenced by movies or screens

  # ── Movies ──────────────────────────────────────────────
  /movies:
    get:
      tags: [Movies]
      summary: Get movies by city with optional filters
      operationId: getMovies
      parameters:
        - in: query
          name: cityId
          required: true
          description: City ID
          schema:
            type: integer
          example: 1
        - in: query
          name: langId
          required: false
          description: Filter movies by language
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: List of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieListResponse'
        '400':
          description: Invalid request
        '500':
          description: Internal server error
    post:
      tags: [Movies]
      summary: Create a new movie (admin)
      operationId: createMovie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieCreateRequest'
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetails'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /movies/{movieId}:
    get:
      tags: [Movies]
      summary: Get full details of a movie
      operationId: getMovieDetails
      parameters:
        - in: path
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        - in: query
          name: cityId
          required: false
          description: If provided, includes availability info for this city
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Movie details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetails'
        '404':
          description: Movie not found
        '500':
          description: Internal server error
    put:
      tags: [Movies]
      summary: Update a movie (admin)
      operationId: updateMovie
      parameters:
        - in: path
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieUpdateRequest'
      responses:
        '200':
          description: Movie updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetails'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Movie not found
    delete:
      tags: [Movies]
      summary: Delete a movie (admin)
      operationId: deleteMovie
      parameters:
        - in: path
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
      responses:
        '204':
          description: Movie deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Movie not found
        '409':
          description: Movie still has shows scheduled
  /movies/{movieId}/language-formats:
    post:
      tags: [Movies]
      summary: Add a language-format combination to a movie (admin)
      operationId: addMovieLanguageFormat
      parameters:
        - in: path
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieLanguageFormatCreateRequest'
      responses:
        '201':
          description: Language-format added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieLanguageFormat'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Movie, language, or format not found
        '409':
          description: Language-format combination already exists for this movie
  /movies/{movieId}/language-formats/{mlfId}:
    delete:
      tags: [Movies]
      summary: Remove a language-format combination from a movie (admin)
      operationId: removeMovieLanguageFormat
      parameters:
        - in: path
          name: movieId
          required: true
          description: Unique movie ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        - in: path
          name: mlfId
          required: true
          description: Unique movie-language-format ID
          schema:
            type: string
            format: uuid
          example: "00000000-0000-4000-c000-000000000001"
      responses:
        '204':
          description: Language-format removed
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Movie or language-format not found
        '409':
          description: Language-format still has shows scheduled

  # ── Shows ───────────────────────────────────────────────
...
components:
  schemas:
...
    CityListResponse:
      type: object
      properties:
        cities:
          type: array
          items:
            $ref: '#/components/schemas/City'
    MovieListResponse:
      type: object
      properties:
        movies:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
...
    City:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Mumbai"
        state:
          type: string
          example: "Maharashtra"
        country:
          type: string
          example: "India"

    # ── Movie ───────────────────────────────────────────────
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        title:
          type: string
          example: Pushpa 2
        genre:
          type: string
          example: Action
        duration:
          type: integer
          example: 180
        rating:
          type: integer
          description: "Rating as a percentage (0-100)"
          minimum: 0
          maximum: 100
          example: 84
    MovieDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        title:
          type: string
          example: Pushpa 2
        description:
          type: string
          example: Action thriller movie
        genre:
          type: string
          example: Action
        duration:
          type: integer
          example: 180
        rating:
          type: integer
          description: "Rating as a percentage (0-100)"
          minimum: 0
          maximum: 100
          example: 84
        releaseDate:
          type: string
          format: date
          example: 2026-02-20
        releasedLanguages:
          type: array
          description: All languages this movie is released in
          items:
            type: string
          example: [Hindi, Telugu, Tamil]
        releasedFormats:
          type: array
          description: All formats this movie is released in
          items:
            type: string
          example: [2D, 3D, IMAX]
        cityLanguageFormats:
          type: array
          description: Languages and their available formats in the requested city. Null if no city was specified.
          items:
            $ref: '#/components/schemas/LanguageFormat'
        cast:
          type: array
          items:
            type: string
          example: [Allu Arjun, Rashmika Mandanna]
        director:
          type: string
          example: Sukumar
        trailerUrl:
          type: string
          example: https://youtube.com/trailer
        posterUrl:
          type: string
          example: https://cdn.movie.com/poster.jpg
        availableInCity:
          type: boolean
          description: Whether the movie is currently running in the requested city. Null if no city was specified in the query.
          example: true

    # ── Show ────────────────────────────────────────────────
...
    LanguageFormat:
      type: object
      properties:
        language:
          type: string
          example: Hindi
        formats:
          type: array
          items:
            type: string
          example: [2D, 3D, IMAX]

    # ── Theatre ─────────────────────────────────────────────
...
    CityCreateRequest:
      type: object
      required:
        - name
        - state
        - country
      properties:
        name:
          type: string
          minLength: 1
          example: Mumbai
        state:
          type: string
          minLength: 1
          example: Maharashtra
        country:
          type: string
          minLength: 1
          example: India
        latitude:
          type: number
          format: double
          example: 19.0760
        longitude:
          type: number
          format: double
          example: 72.8777
    CityUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          example: Mumbai
        state:
          type: string
          minLength: 1
          example: Maharashtra
        country:
          type: string
          minLength: 1
          example: India
        latitude:
          type: number
          format: double
          example: 19.0760
        longitude:
          type: number
          format: double
          example: 72.8777

    # ── Languages ───────────────────────────────────────────
    Language:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Hindi
        code:
          type: string
          example: hi
    LanguageListResponse:
      type: object
      properties:
        languages:
          type: array
          items:
            $ref: '#/components/schemas/Language'
    LanguageCreateRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          minLength: 1
          example: Hindi
        code:
          type: string
          minLength: 1
          maxLength: 8
          description: ISO 639-1 / 639-2 language code
          example: hi
    LanguageUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          example: Hindi
        code:
          type: string
          minLength: 1
          maxLength: 8
          example: hi

    # ── Formats ─────────────────────────────────────────────
    Format:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 2D
    FormatListResponse:
      type: object
      properties:
        formats:
          type: array
          items:
            $ref: '#/components/schemas/Format'
    FormatCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: IMAX
    FormatUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        name:
          type: string
          minLength: 1
          example: IMAX

    # ── Admin: Movies ───────────────────────────────────────
    MovieCreateRequest:
      type: object
      required:
        - title
        - durationMin
        - genre
      properties:
        title:
          type: string
          minLength: 1
          example: Pushpa 2
        description:
          type: string
          example: Action thriller movie
        durationMin:
          type: integer
          minimum: 1
          example: 180
        genre:
          type: string
          minLength: 1
          example: Action
        rating:
          type: integer
          description: Rating as a percentage (0-100)
          minimum: 0
          maximum: 100
          example: 84
        releaseDate:
          type: string
          format: date
          example: 2026-02-20
        posterUrl:
          type: string
          example: https://cdn.movie.com/poster.jpg
        cast:
          type: array
          items:
            type: string
          example: [Allu Arjun, Rashmika Mandanna]
        director:
          type: string
          example: Sukumar
        trailerUrl:
          type: string
          example: https://youtube.com/trailer
    MovieUpdateRequest:
      type: object
      description: All fields optional; only provided fields are updated.
      properties:
        title:
          type: string
          minLength: 1
          example: Pushpa 2
        description:
          type: string
          example: Action thriller movie
        durationMin:
          type: integer
          minimum: 1
          example: 180
        genre:
          type: string
          minLength: 1
          example: Action
        rating:
          type: integer
          minimum: 0
          maximum: 100
          example: 84
        releaseDate:
          type: string
          format: date
          example: 2026-02-20
        posterUrl:
          type: string
          example: https://cdn.movie.com/poster.jpg
        cast:
          type: array
          items:
            type: string
          example: [Allu Arjun, Rashmika Mandanna]
        director:
          type: string
          example: Sukumar
        trailerUrl:
          type: string
          example: https://youtube.com/trailer

    # ── Admin: Movie ↔ Language/Format ──────────────────────
    MovieLanguageFormat:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "00000000-0000-4000-c000-000000000001"
        movieId:
          type: string
          format: uuid
          example: "00000000-0000-4000-a000-000000000001"
        languageId:
          type: integer
          example: 1
        languageName:
          type: string
          example: Hindi
        formatId:
          type: integer
          example: 2
        formatName:
          type: string
          example: 3D
    MovieLanguageFormatCreateRequest:
      type: object
      required:
        - languageId
        - formatId
      properties:
        languageId:
          type: integer
          minimum: 1
          example: 1
        formatId:
          type: integer
          minimum: 1
          example: 2

    # ── Admin: Theatres ─────────────────────────────────────
...